+++ frontend/src/app/magic/magic-card-list/magic-card-modal/magic-card-modal.component.ts
@@ -18,7 +18,7 @@ export class MagicCardModalComponent implements OnInit, AfterViewInit {
 
     isLoggedIn!: boolean;
     otherVersionCards?: Card[];
-    allVerions!: Card[];
+    allVerions?: Card[];
 
     constructor(
         private authenticationService: AuthenticationService,
@@ -56,6 +56,7 @@ export class MagicCardModalComponent implements OnInit, AfterViewInit {
         const nextMagicCard = this.magicCardModalService.getNextCard();
         if (nextMagicCard) {
             this.magicCard = nextMagicCard;
+            this.allVerions = undefined;
             this.otherVersionCards = undefined;
         }
     }
@@ -64,6 +65,7 @@ export class MagicCardModalComponent implements OnInit, AfterViewInit {
         const nextMagicCard = this.magicCardModalService.getPreviousCard();
         if (nextMagicCard) {
             this.magicCard = nextMagicCard;
+            this.allVerions = undefined;
             this.otherVersionCards = undefined;
         }
     }
@@ -110,12 +112,12 @@ export class MagicCardModalComponent implements OnInit, AfterViewInit {
 
     onChangeVersion(changeVersion: Card) {
         this.magicCard = changeVersion;
-        this.otherVersionCards = this.allVerions.filter(
-            card =>
-                !(
-                    card.cardExpansion === this.magicCard.cardExpansion &&
-                    card.cardNumber === this.magicCard.cardNumber
-                ),
-        );
+        // this.otherVersionCards = this.allVerions.filter(
+        //     card =>
+        //         !(
+        //             card.cardExpansion === this.magicCard.cardExpansion &&
+        //             card.cardNumber === this.magicCard.cardNumber
+        //         ),
+        // );
     }
 }