+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added Vow cards
 - Added Dmu cards
 
+### Changed
+
+- New style in card modal for card versions
+
 ## [1.5.0] - 2021-11-08
 
 ### Added